Install requirements before copying files
We want to install the apt (etc.) requirements in their own step before we copy in files. This is so that we can cache the result of installing the requirements. This should significantly speed up development, since we won't need to wait for all the apt installs every time we make a change and rebuild the container.