Find package when building
A module is when we have the python file in the root of the repo. This worked in the Docker container because we copied the contents of the flask_opensearch directory into the root. However, it doesn't work in the release GitHub action because there is no flask_opensearch.py in the root of the repo. So the current version released to pypi.org contains no code.

Instead, what we need to use is packages (https://setuptools.pypa.io/en/latest/deprecated/distutils/examples.html?highlight=py_modules#pure-python-distribution-by-package). This ensures that `python build` can pick up the package.

Update the Dockerfile to match.